refactor: /community/[boardCode]/[postId] 페이지 전환
